refactor: 아토믹 디자인 시스템과 혼동을 없애기 위해 components/common을 components/utils로 변경
